backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["araki777 <arakichi777@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
requests = "^2.32.3"
websockets = "^14.2"
python-dotenv = "^1.0.1"
httpx = "^0.28.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"